requirements.txt
appdirs == 1.4.4
asgiref == 3.2.10
distlib == 0.3.1
Django==3.1.1
filelock == 3.0.12
gunicorn == 20.0.4
Pillow == 7.2.0
pytz == 2020.1
six == 1.15.0
sqlparse == 0.3.1
whitenoise == 5.1.0
dj-database-url == 0.5.0
psycopg2 == 2.8.5
social-auth-app-django == 4.0.0
django-bootstrap3 == 14.1.0
django-rest-framework-social-oauth2 == 1.1.0


